# .history/api/swagger/swagger_20200109160346.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Recipe App
# during dev, should point to your local machine
host: localhost:4000
# basePath prefixes all resource paths 
basePath: /v1  
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
  - text/plain
paths:
  /recipes:
    x-swagger-router-controller: recipie
    get:
      description: Return all the recipes
      operationId: getRecipes
      responses: 
        "200":
          description: Sucess get all the recipes
          schema:
            type: array
            items:
              $ref: "#/definitions/Recipe"
        "500":
          description: Unexpected Error
          schema:
            type: object
            properties:
              message:
                type: string
    post:
      description: Create one new Recipe
      operationId: postRecipes
      parameters:
        - in: body
          name: body
          description: The recipe to be added
          required: true
          schema:
            $ref: "#/definitions/Recipe"
      responses:
        "204":
          description: Sucess
        "500":
          description: unexpected Error
          schema:
            type: object
            properties:
              message:
                type: string
      
  /recipes/{recipeId}:
    x-swagger-router-controller: recipe
    parameters:
    - in: path
      name: recipeId
      type: integer
      required: true
      description: Numeric Id of the recipe
    delete:
      description: Remove a recipe
      operationId: deleteRecipe
      responses:
        "204":
          description: Success
        "404":
          description: The Id passed does not match any recipe.
        "500":
          description: unexpected Error
          schema:
            type: object
            properties:
                message:
                    type: string
    put:
      parameters:
      - in: body
        name: Recipe
        schema:
          $ref: "#/definitions/Recipe"
        required: true
        description: Id of the recipe you want to get the comments.
      description: updates a Recipe
      operationId: putRecipes
      responses:
        "204":
          description: Success
        "404":
          description: The Id passed does not match any recipe.
        "500":
          description: unexpected Error
          schema:
            type: object
            properties:
                message:
                    type: string
                    
  /comments:
    x-swagger-router-controller: comments
    get:
      description: Return all the comments
      operationId: getComments
      responses: 
        "200":
          description: Sucess get all the comments
          schema:
            type: array
            items:
              $ref: "#/definitions/Comment"
        "500":
          description: Unexpected Error
          schema:
            type: object
            properties:
              message:
                type: string
    post:
      description: Create one new Comment
      operationId: postComment
      parameters:
          - in: body
            name: body
            description: The recipe to be added
            required: true
            schema:
              $ref: "#/definitions/Recipe"
      responses:
        "204":
            description: Sucess
        "500":
            description: unexpected Error
            schema:
              type: object
              properties:
                message:
                  type: string             
  /comments/{commentId}:
    x-swagger-router-controller: comment
    parameters:
    - in: path
      name: commentId
      type: integer
      required: true
      description: Numeric Id of the recipe
    delete:
      description: Remove a Comment
      operationId: deleteComment
      responses:
        "204":
          description: Success
        "404":
          description: The Id passed does not match any comment.
        "500":
          description: unexpected Error
          schema:
            type: object
            properties:
                message:
                    type: string

  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  Recipe:
    properties:
      id:
        type: integer
      isPublic:
        type: boolean
      name: 
        type: string
      chef:
        type: string
        enum: [Joao, Charles]
      stepList:
        type: array
        items:
          type: string
      ingredients:
        type: array
        items:
          type: object
          properties:
            amount:
              type: number
            name:
              type: string
  Comment:
    properties:
      id:
        type: integer
      recipieId:
        type: integer
      author: 
        type: string
      text:
        type: string
      date:
        type: string             